Move MountSearchableActionRequest and Action to xpack/core
This moves the classes needed to invoke the mount snapshot request to the
xpack/core module such that clients aiming to invoke the mount api don't need
to add a dependency on the xpack/searchable-snapshots module.
